Move utxo validation functions to a new virtual processor mod
