+++ consensus/src/pipeline/virtual_processor/mod.rs
@@ -1,2 +1,3 @@
 mod processor;
+mod utxo_validation;
 pub use processor::*;
