+++ consensus/src/pipeline/virtual_processor/processor.rs
@@ -5,7 +5,7 @@ use crate::{
         services::reachability::{MTReachabilityService, ReachabilityService},
         stores::{
             block_transactions::{BlockTransactionsStoreReader, DbBlockTransactionsStore},
-            ghostdag::{DbGhostdagStore, GhostdagData, GhostdagStoreReader},
+            ghostdag::{DbGhostdagStore, GhostdagStoreReader},
             headers::{DbHeadersStore, HeaderStoreReader},
             reachability::DbReachabilityStore,
             statuses::{
@@ -22,24 +22,18 @@ use crate::{
 use consensus_core::{
     block::Block,
     blockhash,
-    header::Header,
     muhash::MuHashExtensions,
-    tx::{PopulatedTransaction, Transaction, ValidatedTransaction},
-    utxo::{
-        utxo_collection::UtxoCollection,
-        utxo_collection::UtxoCollectionExtensions,
-        utxo_diff::UtxoDiff,
-        utxo_view::{self, UtxoView},
-    },
+    tx::ValidatedTransaction,
+    utxo::{utxo_collection::UtxoCollection, utxo_collection::UtxoCollectionExtensions, utxo_diff::UtxoDiff, utxo_view},
     BlockHashMap, BlockHashSet,
 };
-use crossbeam_channel::Receiver;
 use hashes::Hash;
-use itertools::Itertools;
 use kaspa_core::trace;
 use muhash::MuHash;
+
+use crossbeam_channel::Receiver;
+use itertools::Itertools;
 use parking_lot::RwLock;
-use rayon::prelude::*;
 use std::{
     collections::hash_map::Entry::{Occupied, Vacant},
     iter::FromIterator,
@@ -268,105 +262,6 @@ impl VirtualStateProcessor {
             Ok(())
         } else {
             Ok(())
-        }
-    }
-
-    /// Verify that the current block fully respects its own UTXO view. We define a block as
-    /// UTXO valid if all the following conditions hold:
-    ///     1. The block header includes the expected `utxo_commitment`.
-    ///     2. The block header includes the expected `accepted_id_merkle_root`.
-    ///     3. The coinbase transaction rewards the mergeset blocks correctly.
-    ///     4. All block transactions are valid against its own UTXO view.
-    fn verify_utxo_validness_requirements<V: UtxoView + Sync>(
-        self: &Arc<VirtualStateProcessor>,
-        utxo_view: &V,
-        header: &Header,
-        mergeset_data: &GhostdagData,
-        multiset_hash: &mut MuHash,
-        mut accepted_tx_ids: Vec<Hash>,
-        mergeset_fees: BlockHashMap<u64>,
-    ) -> BlockStatus {
-        // Verify header UTXO commitment
-        let expected_commitment = multiset_hash.finalize();
-        if expected_commitment != header.utxo_commitment {
-            trace!("wrong commitment: {}, {}, {}", header.hash, expected_commitment, header.utxo_commitment);
-            return StatusDisqualifiedFromChain;
-        } else {
-            trace!("correct commitment: {}, {}", header.hash, expected_commitment);
-        }
-
-        // Verify header accepted_id_merkle_root
-        accepted_tx_ids.sort();
-        let expected_accepted_id_merkle_root = merkle::calc_merkle_root(accepted_tx_ids.iter().copied());
-        if expected_accepted_id_merkle_root != header.accepted_id_merkle_root {
-            trace!("wrong accepted_id_merkle_root: {}, {}", expected_accepted_id_merkle_root, header.accepted_id_merkle_root);
-            return StatusDisqualifiedFromChain;
-        }
-
-        let txs = self.block_transactions_store.get(header.hash).unwrap();
-        let coinbase = &txs[0];
-
-        // Verify coinbase transaction
-        // TODO: verify coinbase using `mergeset_fees`
-        // TODO: build expected coinbase
-
-        // Verify all transactions are valid in context (TODO: skip validation when becoming selected parent)
-        let validated_transactions = self.validate_transactions_in_parallel(&txs, &utxo_view, header.hash, header.hash);
-        if validated_transactions.len() < txs.len() - 1 {
-            // Some transactions were invalid
-            return StatusDisqualifiedFromChain;
-        }
-
-        StatusUTXOValid
-    }
-
-    /// Validates transactions against the provided `utxo_view` and returns a vector with all transactions
-    /// which passed the validation
-    fn validate_transactions_in_parallel<'a, V: UtxoView + Sync>(
-        self: &Arc<VirtualStateProcessor>,
-        txs: &'a Vec<Transaction>,
-        utxo_view: &V,
-        merged_block: Hash,
-        merging_block: Hash,
-    ) -> Vec<ValidatedTransaction<'a>> {
-        txs
-            .par_iter() // We can do this in parallel without complications since block body validation already ensured 
-                        // that all txs in the block are independent 
-            .skip(1) // Skip the coinbase tx. 
-            .filter_map(|tx| self.validate_transaction_in_utxo_context(tx, &utxo_view, merged_block, merging_block))
-            .collect()
-    }
-
-    /// Attempts to populate the transaction with UTXO entries and performs all tx validations
-    fn validate_transaction_in_utxo_context<'a>(
-        self: &Arc<VirtualStateProcessor>,
-        transaction: &'a Transaction,
-        utxo_view: &impl UtxoView,
-        merged_block: Hash,
-        merging_block: Hash,
-    ) -> Option<ValidatedTransaction<'a>> {
-        let mut entries = Vec::with_capacity(transaction.inputs.len());
-        for input in transaction.inputs.iter() {
-            if let Some(entry) = utxo_view.get(&input.previous_outpoint) {
-                entries.push(entry.clone());
-            } else {
-                trace!("missing entry for block {} and outpoint {}", merged_block, input.previous_outpoint);
-                break;
-            }
-        }
-        if entries.len() < transaction.inputs.len() {
-            // Missing inputs
-            return None;
-        }
-        let populated_tx = PopulatedTransaction::new(transaction, entries);
-        let res = self.transaction_validator.validate_populated_transaction_and_get_fee(&populated_tx, merging_block);
-        // TODO: pass DAA score instead of hash to function above ^^
-        match res {
-            Ok(calculated_fee) => Some(populated_tx.to_validated(calculated_fee)), // TODO: collect fee info and verify coinbase transaction of `chain_block`
-            Err(tx_rule_error) => {
-                trace!("tx rule error {} for block {} and tx {}", tx_rule_error, merged_block, transaction.id());
-                None // TODO: add to acceptance data as unaccepted tx
-            }
         }
     }
